rag/requirements.txt
# RAG System Dependencies
langchain==0.1.10
langchain-community==0.0.24
chromadb==0.4.22
openai==1.12.0
tiktoken==0.6.0
anthropic==0.18.1
python-dotenv==1.0.1
watchdog==4.0.0
